requirements.txt
flask==3.0.0
pymysql==1.1.0
pyjwt==2.8.0
werkzeug==3.0.1
cryptography==41.0.7